backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
chromadb==0.5.5
pdfplumber==0.11.4
langchain-text-splitters==0.2.2
sentence-transformers==3.0.1
google-generativeai==0.5.4
pytest==8.3.2


